+++ package.json
@@ -28,6 +28,7 @@
     "express-session": "^1.17.2",
     "https": "^1.0.0",
     "jsonwebtoken": "^8.5.1",
+    "mysql": "^2.18.1",
     "nodemon": "^2.0.15",
     "pug": "^3.0.2",
     "qs": "^6.10.3",
