+++ views/index_test.html
@@ -1,40 +1,40 @@
 <!DOCTYPE html>
 <html lang="ko">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Test Social Login</title>
-    </head>
-    <body>
-        <!-- 카카오 로그인 -->
-       <H1>카카오 로그인</H1>
-       <a href="javascript:kakaoLogin();">카카오 로그인</a>
-       <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
-       <script>
-           //SDK를 초기화 할 자바스크립트 Key
-           window.Kakao.init("e203110d85291da30da407c044fc5d25");
-           //SDK 초기화 여부, True: 정상, False: 키가 올바르지 않음
-           console.log(Kakao.isInitialized());
-           function kakaoLogin() {
-              // 1.함수가 로그인을 시도할때 
-               window.Kakao.Auth.login({
-                //우리가 개인 정보 허용범위를 Scope안에 작성
-                scope: 'profile_nickname, profile_image, account_email',
-                //로그인 성공 시
-                success: function(auth_obj) {
-                    console.log(auth_obj)
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Test Social Login</title>
+  </head>
+  <body>
+    <!-- 카카오 로그인 -->
+    <h1>카카오 로그인</h1>
+    <a href="javascript:kakaoLogin();">카카오 로그인</a>
+    <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
+    <script>
+      //SDK를 초기화 할 자바스크립트 Key
+      window.Kakao.init('e203110d85291da30da407c044fc5d25');
+      //SDK 초기화 여부, True: 정상, False: 키가 올바르지 않음
+      console.log(Kakao.isInitialized());
+      function kakaoLogin() {
+        // 1.함수가 로그인을 시도할때
+        window.Kakao.Auth.login({
+          //개발자가 허용한 개인 정보를 Scope안에 작성
+          scope: 'profile_nickname, profile_image, account_email',
+          //로그인 성공 시
+          success: function (auth_obj) {
+            console.log(auth_obj);
 
-                    const reqObj = {
-                        url: '/v2/user/me',
-                        success: res => {
-                            const kakao_account = res.kakao_account;
-                            console.log(kakao_account)
-                        }
-                    }
-                    window.Kakao.API.request(reqObj);
-                }
-            })
-           }
-       </script>
-    </body>
+            const reqObj = {
+              url: '/v2/user/me',
+              success: (res) => {
+                const kakao_account = res.kakao_account;
+                console.log(kakao_account);
+              },
+            };
+            window.Kakao.API.request(reqObj);
+          },
+        });
+      }
+    </script>
+  </body>
 </html>
