+++ views/pages/index.ejs
@@ -12,7 +12,7 @@
     <!-- 카카오 로그인 -->
     <H1>카카오 로그인</H1>
     <a href="/oauth/kakao">카카오 로그인</a>
-    <a href="/oauth/kakao/account_info">오도짜세프로필</a>
+    <a href="/oauth/kakao/account_info">프로필</a>
     <a href="/oauth/kakao/logout">로그아웃</a>
     <a href="/oauth/kakao/unlink">인증 삭제</a>
   </body>
